2/12 Experience memory
Add a memory space for the model, but not yet implement the learning
routine.
Change the output of the network to 2 outputs indicating turning left or
right.  It's wierd to use only 1 output with Q-learning.
Push the model file back with the rests of the file. But maybe requires
a better file structure if we're going to implement more models.